Rename files for intuitiveness
