+++ python/ray/llm/tests/serve/cpu/common/test_models_common.py
@@ -1,6 +1,6 @@
 import pytest
 
-from ray.llm._internal.serve.configs.rayllm_models import ImageContent, Message, TextContent
+from ray.llm._internal.serve.configs.prompt_formats import ImageContent, Message, TextContent
 
 
 def test_validation_message():
